Moved the only remaining function from DataTypeConversions to DateUtils
